+++ DustyPig.TVDB/Models/TagOption.cs
@@ -4,11 +4,11 @@ namespace DustyPig.TVDB.Models
     {
         public string HelpText { get; set; }
 
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         public string Name { get; set; }
 
-        public int Tag { get; set; }
+        public long Tag { get; set; }
 
         public string TagName { get; set; }
     }
